Version 0.2
This version now supports a new type of close, the "middle close". For XML/SGML, it will close the tag and reopen the same tag (useful for list of tags); for LaTeX, it will insert \item inside of {itemize}, {enumerate}, .. environments, just close otherwise.

Middle closes are invoke by the two-key sequence CTRL-\ =